add warning about time scope of these credentials
I tried to see if these credentials could have a longer duration. The
endpoint we're hitting to get them is undocumented but similar to other
assume-role endpoints. I tried adding a `DurationSeconds` param to the
url but it was ignored. Any extra params I added to the URL were
ignored. Guessing this endpoint is hard-coded to ignore anything passed
in and always return 15min long sessions

Signed-off-by: Taylor Silva <tsilva@prophix.com>